requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
dj-database-url==0.5.0
Django==4.0.2
django-cors-headers==3.11.0
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.3
inflection==0.5.1
inscriptis==2.2.0
itypes==1.2.0
Jinja2==3.0.3
lxml==4.8.0
MarkupSafe==2.1.0
openapi-codec==1.3.2
packaging==21.3
pdfkit==1.0.0
Pillow==9.0.1
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
PyPDF2==1.26.0
python-bidi==0.4.2
python-decouple==3.6
pytz==2021.3
PyYAML==6.0
reportlab==3.6.6
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
webencodings==0.5.1
whitenoise==6.0.0
xhtml2pdf==0.2.5
